move images to public
