+++ public/graphics/onboarding-border.png
