+++ src/components/OnboardingManager.tsx
@@ -1,15 +1,15 @@
 'use client';
 
 import { FormContainer, LogoImage, OuterDiv } from '@/app/onboarding/styles';
-import IJPLogoBlue from '@/assets/images/ijp_logo_blue.webp';
 import CONFIG from '@/lib/configs';
 import { useAuth } from '@/utils/AuthProvider';
 import { useGuardedOnboarding, useOnboardingNavigation } from '@/utils/hooks';
 import { useProfile } from '@/utils/ProfileProvider';
+import IJPLogoBlue from '~/public/images/ijp_logo_blue.webp';
 import { useRouter } from 'next/navigation';
 import { ReactNode, useEffect } from 'react';
-import ProgressBar from './ProgressBar';
 import { UnstyledButton } from './Buttons';
+import ProgressBar from './ProgressBar';
 
 export default function OnboardingManager({
   children,
