+++ src/components/ProgressBar/styles.ts
@@ -2,7 +2,7 @@ import COLORS from '@/styles/colors';
 import { H4 } from '@/styles/text';
 import Link from 'next/link';
 import styled from 'styled-components';
-import dashedBorder from '@/assets/graphics/onboarding-border.png';
+import dashedBorder from '../../../public/graphics/onboarding-border.png';
 
 export const ProgressBarContainer = styled.div<{ $show: boolean }>`
   display: ${({ $show }) => ($show ? 'grid' : 'none')};
